Tidying up CAN drivers (part 2)
